Locally
An Android App that helps in keeping details on a hold for a while. Why to go through contacs and save.Here,enter the details and save.Afterwards,if neccessary,transfer it to contacs and save directly.
I have used Shared Preferences for storing locally and on a hold.
Help is always  appreciated.