A Local App Book which allows to store temporary contacts for a Temporary time period
Added features like Direcct calling on a Number from the app by requestiong permissions.Also transferring the number to the dial pad and phonebook has been added with Menu icon in the Next Activity Page.Features like Rating and Settings page are being currently worked on. Contributions are always welcomed and suggestions too, to improve the experience of the App.